requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
matplotlib.pyplot
seaborn
